part 4 completed

+++ part4/bloglist_backend/controllers/blogs.js
@@ -1,37 +1,59 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response, next) => {
-  const body = request.body;
+blogsRouter.post(
+  "/",
+  middleware.userExtractor,
+  async (request, response, next) => {
+    const body = request.body;
+    const user = request.user;
 
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-  });
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user.id,
+    });
 
-  try {
-    const savedBlog = await blog.save();
-    response.status(201).json(savedBlog);
-  } catch (exception) {
-    next(exception);
+    try {
+      const savedBlog = await blog.save();
+      user.blogs = user.blogs.concat(savedBlog._id);
+      await user.save();
+      response.status(201).json(savedBlog);
+    } catch (exception) {
+      next(exception);
+    }
   }
-});
+);
 
-blogsRouter.delete("/:id", async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
-  } catch (exception) {
-    next(exception);
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response, next) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog && user._id.toString() === blog.user._id.toString())
+      try {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+      } catch (exception) {
+        next(exception);
+      }
+    else
+      return response.status(401).json({ error: `Unauthorized: invalid user` });
   }
-});
+);
 
 blogsRouter.put("/:id", async (request, response, next) => {
   const body = request.body;
